refactor: rework types to avoid unwanted anys
