+++ packages/core/components/Render/index.tsx
@@ -4,9 +4,13 @@ import { rootDroppableId } from "../../lib/root-droppable-id";
 import { Config, Data } from "../../types/Config";
 import { DropZone, DropZoneProvider } from "../DropZone";
 
-export function Render<
-  UserConfig extends Config<any, any, any> = Config<any, any, any>
->({ config, data }: { config: UserConfig; data: Data }) {
+export function Render<UserConfig extends Config = Config>({
+  config,
+  data,
+}: {
+  config: UserConfig;
+  data: Data;
+}) {
   // DEPRECATED
   const rootProps = data.root.props || data.root;
   const title = rootProps?.title || "";
