+++ packages/core/components/ServerRender/index.tsx
@@ -59,9 +59,13 @@ function DropZoneRender({
   );
 }
 
-export function Render<
-  UserConfig extends Config<any, any, any> = Config<any, any, any>
->({ config, data }: { config: UserConfig; data: Data }) {
+export function Render<UserConfig extends Config = Config>({
+  config,
+  data,
+}: {
+  config: UserConfig;
+  data: Data;
+}) {
   if (config.root?.render) {
     // DEPRECATED
     const rootProps = data.root.props || data.root;
